Fixes #1
Fixes a problem where a remote disconnect would not immediately end the transfer on the server side if a drop site was hanging because of an error.

This also fixes the equivalent problem if the transfer completed successfully while a drop site was hanging because of an error. In that case, a file transfer would not end on the server side until all error timeouts had elapsed.